Some refactoring. Now: one layer threadpool. But decide to revert.
Now one layer thread pool works, but no sigificant acceleration.
According to some experiments, it looks so fine.
The reason why it is not fast, may be that network is not slow,
and creating threads takes time.

However, the code looks ugly now.

For the purpose of clearance, I decide to revert it to a previous version:

commit 86c4727b14ef8cbb345eb636a8cd3d230b53fc2c
Author: gn01842919 <gn01842919@gmail.com>
Date:   Sun Feb 25 15:33:51 2018 +0800

Modify the concurrency algorithm.
Processing time for local sources reduced significantly, be total time remains.